More 64bit fixes

+++ xLights/outputs/ArtNetOutput.cpp
@@ -300,8 +300,8 @@ std::string ArtNetOutput::GetLongDescription() const
 
     if (!_enabled) res += "INACTIVE ";
     res += "ArtNet " + _ip + " {" + GetUniverseString() + "} ";
-    res += "[1-" + std::string(wxString::Format(wxT("%i"), _channels)) + "] ";
-    res += "(" + std::string(wxString::Format(wxT("%i"), GetStartChannel())) + "-" + std::string(wxString::Format(wxT("%i"), GetEndChannel())) + ") ";
+    res += "[1-" + std::string(wxString::Format(wxT("%i"), (long)_channels)) + "] ";
+    res += "(" + std::string(wxString::Format(wxT("%i"), (long)GetStartChannel())) + "-" + std::string(wxString::Format(wxT("%i"), (long)GetEndChannel())) + ") ";
     res += _description;
 
     return res;
